Refactored main.py to be much smaller 750 down to 250 lines and it doesn't crash on updates.
